Style Update
Updated the styling to be more consistent, align the header more, as well as add a fair bit of spacing and color coordination. I also added in a test version of adding a new button. This needs to be rewritten to be much more modular, but for now you can add a new button based on having earned a progression node. This should instead be implemented as a 2nd set of progression nodes and just iterate through that list, rather than currently I jammed it into the pointer progression.