Progression Update and First Animation
Big progression update!

First we restructured and renamed a variety of files. Now we have progression for the overall garden, individual plants, as well as a progression tree for specifically unlocking different paths (and a way to check those nodes to thus unlock content), and one for user interface upgrade for the plants. This is so that you can opt into stuff if you want to, but you dont have to. I should also add in a settings panel but we're not there yet. But there is for instance a progression tree for Carrot, and that will unlock multipliers for the carrot, but seeing some of those nodes is locked behind unlocking some user interface nodes. I should also make some of those nodes unlocked based on achieving certain things, and also track totals of Kors earned etc.

The other big thing with this pass was updating the visuals for the progression side of the screen. Now there is a bar with different progression trees as tabs, although it still needs more work. Namely that it should be solid I think, and also all plant progression trees should be moved to a single drop down, that was the bar doesnt over crowd and we have to make it scroll.

I think there was also some work done during this update to make coloring and styling more consistent across the site. I dont remember off hand.

The other big update as getting our first animation going. It will be the Kors Icon I think once we clean it up a bit, make it a re-usable component, and also play with the colors some more. it's a glowing rotating group of squares!!

Also with this we now  have a real path forward with styling. The next styling update will probably be a full pass over the site, making it full neon-y and all the components have fun animations etc.

Still a lot of work to do, but good progress none the less!