Clicker Update and Kors Icon Update
Started by updating the Kors Icon we made last update to be dynamic so it can be reused and resized automatically without having to do all the math every time.

The next big thing from this update was updating the clickers. Instead of just being a bit of text we now have a full grow cycle and some chunky buttons with better text.

Updated the ProgressionType to be a string enum for easy printouts of plant names.

Made the title bar better? at least now it has the shimmer. that being said need to QA the shimmer because it might be nauseating. either make it move differently I think or just be static. TBD.

Todo:

1. Buttons need some work after the console is taken off, they are larger than they were in testing.
2. Kors per click is down by 5% so we may need to start with a higher garden multiplier? that would prob fix it tbh.,
3. NEED to update the way progression nodes look before and after purchase.
4. maybe move the tech buttons / add new ones.
5. Actual need to do: add animations to all the things that give the interactions life. maybe sounds too, but at the very least I need more like, reaction from the UI when I hover / click / etc.
6. Add in all the progressions.